Move Qt code for dendrogram viewer into qt sub-directory
