+++ glue/plugins/dendro_viewer/__init__.py
@@ -1,5 +1,5 @@
 def setup():
     from glue.config import qt_client
-    from .qt_widget import DendroWidget
+    from .qt.viewer_widget import DendroWidget
     from .data_factory import load_dendro
     qt_client.add(DendroWidget)
