+++ glue/plugins/dendro_viewer/qt/viewer_widget.py
